refactor: move get_new_package_ref/get_new_upgrade_cap_ref out of iota-json-rpc-types
